Adding generic GETDATA and GETCOLUMN proxy methods for _uuid and _version in TypedBaseTable and enhanced the IT to monitor on OpenVSwitch Table
Please note that as per RFC 7047, the _uuid column will not be included for filter-less monitors and hence getUuid will not work if we are monitoring with no Filters.
But the TableUpdate contains the UUID for each of the update and a monitoring solution should manage that.
Please refer to TypedVSwitchdSchemaIT.java 's getOpenVSwitchTableUuid() for more info on how this is used.

Change-Id: Ia758ed726bbcfdb9fa0b6a7873f50831aed217e4
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>