DB table interfaces for ipfix, netflow and sflow
Patchset5: ugh git add ftw.
Patchset4: more typos ftw
Patchset3: spelling typos

patchset2: These are the ones with IT I have done.
Abondoning the other commits, the groupings are wrong
as some IT tests have dependencies on tables that
I had grouped in another commit.

note: breaking up commits for easier review. next up
is IT for the tables in the new schema.

Change-Id: I8983305659066ee6ab80a166d3ddb0ef55773700
Signed-off-by: Brent Salisbury <brent.salisbury@gmail.com>