Northbound migration to use the new Schema independent Library based Plugin.
Following the Plugin Migration, this commit brings in the Northbound migration.
Due to the nature of the Library rearchiteture, the northbound APIs should not assume any schema.
But unfortunately, that is not the case with the existing APIs and hence we have split the APIs as V2 (older version)
and the newer V3.

This commit just migrates the V2 APIs to the newer Infra. V3 will be added soon.

Also, the GET serailization is still pending and will be addressed in subsequent commit.
But, all the IT tests are passing in this version.

Change-Id: I7c0e59f41499dadb2215fef3d597f98b2efa3c13
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>