Adjusting ovsdb project to work with the new default openflowplugin and openflowjava bundles.
With the latest changes in the Controller and Integration projects, MD-SAL based Openflowplugin is made
the default openflow plugin. In order to work in tandem with Controller/Integration project, this changeset
contains the changes that will make MD-SAL Openflow Plugin as the default Openflow Provider in the OVSDB local
distribution target.

Change-Id: I11fc6bae0fb8c471fc8518baacf3e96eeefce1d1
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>