Bug 1401 : Resolving unrecognized field "syntax" error in OperationResult
As per RFC 7047, the error field is defined only the the "error" and "details" elements.
But it can also contain other elements that is not described by the RFC.

      "An <error> MAY also have other members that describe the error in more detail.
      This document does not specify the names or values of these members."

The ovsdb-server implementation may choose to add more fields to give more information
to the error.

Since, these fields are not defined in the RFC, we should not fail the JSON parsing
if we encounter unknown fields.

Change-Id: I95c2e5ada6d8b97ff6dfe4f4ed9534ba143e822f
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>