Add MD-SAL Adapter for the OVSDB Plugin
This commit adds an initial MD-SAL connector for OVSDB
The OvsdbInventoryManager updates the MD-SAL with nodes as they are
added and removed from inventory. The Yang model provides a reference
between OVSDB nodes and the OpenFlow nodes (bridges) that they manage

Also changes the nodeAdded API in OvsdbIventoryListener to pass the
IpAddress and Port number of the new node. This avoids having to query
the AD-SAL based properties set for this inforamtion

Change-Id: Ia2d8140c3f504d9b2d78b490bb273808f9a86c51
Signed-off-by: Dave Tucker <djt@redhat.com>