Fixing a couple of Integration distribution issues due to the legacy vs MD-SAL openflowplugin conflicts. After the introduction of NX-Extension support, the OVSDB project has moved from a soft-dependancy (via MD-SAL Apis) to a more hard-depedency with Openflowplugin/Openflowjava projects. Due to that reason, the filtering done during the [-of10|-of13] switch for legacy vs MD-SAL plugin selection needs some improvement and that is taken care in the integration project using https://git.opendaylight.org/gerrit/#/c/10003/.
Unfortunately, the net-virt-providers currently carries both OF10, OF13 providers together and hence, filtering that
bundle is not an option based on the above flag switch. Hence, added optional-dependency directive to make sure the
OSGi runtime dont scream when it finds a missing bundle for NX-Extension when running in legacy -of10 mode.

Better fix would have been to reorganize the providers to their own bundles and fitering them out using the above switch.
But since Integration project is impacted, this gives a good smooth transition to the newer model.

Change-Id: I11992a2ba872a0ac49c8a815dc7b4a5bce354b41
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>