Bug 1695 - ignoring pipeline rules for any non-forwarding bridge
By design, the AbstractServiceInstance kept the isBridgeInPipeline as an abstract method for its child concrete Services
to determine the behaviour.

But, the most important and basic case of allowing the pipeline flows on br-int and ignoring on all other Bridge seems
reasonable default in AbstractServiceInstance. This can be overriden by child Services if need be.

In addition to this fix, have also introduced a minor delay in PipelineOrchestrator as soon as a Node event is observed.
This should workaround one of the basic bugs (1997) on the pipeline management case at the minimum.

Change-Id: I34f1d24d3f72695851bda4379483c8ea2b292308
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>