Bug 2004 - Fixing issue where local ports are not receivers for flood originating from another local port
The fix was to update classifier to track whether a flood is coming from local or
remote, and then matching on it to apply the enumerated flood action. The fix also
required OF13Provider to register local ports with the flood rule creation logic.

Reg0 is used by Classifier and LBaaS is moved to Reg1 + Reg2.

Change-Id: I619247349631f909dcc77177bd4cfe733afe7d50
Signed-off-by: Srini Seetharaman <srini.seetharaman@gmail.com>