Bug 1806 - Null destination InetAddress in OF13Provider.handleInterfaceUpdate
The error is caused when ovsdb can not find the local_ip address from the open_vSwitch table other_config row. this address is used to get the src or dst ip address to be used for tunnels.

Added null chcks around the src and dst ip addresses. Log a warning if either is missing.

Change-Id: I57e0c69d699552d5ed07262210e25c8a4555ca66
Signed-off-by: Sam Hague <shague@redhat.com>