Add karaf-based pax-exam so that net-virt can be tested.
Use the following command to run the NetVirtIT test:

mvn -Dit.test=NetVirtIT verify -Pintegrationtest -Dovsdbserver.ipaddress=192.168.120.31 -Dovsdbserver.port=6640 -Pkarafit

This will start the test and bring up the ovsdb karaf distribution. You should first build the distribution and install it to your local m2 repo. When the odl-ovsdb-openstack feature is fully loaded you can then manual configure the bridge to connect to odl with ovs-vsctl set-manager tcp:192.168.120.1:6640.

By default the integration tests exclude NetVirtIt. This is because the current test cannot get service interfaces to make use of their services. An example is getting the connect interfce from the plugin to initiate and outgoing connection to an ovsdb node. The usual "ServiceHelper.getGlobalInstance" did not work in the karaf pax-exam container so another workaround needs to be found.

The test will wait for 10s for the connection from the ovsdb node. Look for the output: "Try to connect!" before connecting. This message indicates the odl-ovsdb-openstack feature is active and ready to receive connections. The timeout can be increased if needed.

Change-Id: I17a944542b564f456f841623083635ac9c4304d0
Signed-off-by: Sam Hague <shague@redhat.com>