Interface ofport and ofport_request support
Added support to yang model for ofport and ofport_request.  These fields are
settable at the time of endpoint creation.  Changed the Interface schema ofport
field to utilize the type Integer; ofport is defined as an integer in the
Open_vSwitch specification.  Changed references to Interface ofport column to
use Integer data type.  Enforced a Yang model range for ofport_request to
ensure the value is in the range [1,65279] as defined by the Open_vSwitch
specification. Future work will persist these settings to the data store.

Change-Id: Ic8f1bab8d8fefe6b102cfceb50ba74f8cfa0b9b7
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>