Workaround to fix southbound it
The problem was the dataBroker in the southbound-it was not right. Something from outside ovsdb changed and caused issues with the way the soutbound-it was getting the broker.

The workaround was to get the broker from the SouthboundProvider.

Also added basic NetworkTopology and Topology to read the OVSDBTopology from the datastores since those values have to exist for southbound-impl to work.

Change-Id: Ie24c2bed42976efc64c142c30f6b1a8935435209
Signed-off-by: Sam Hague <shague@redhat.com>