Fix for BUG 4067 (LBaaS flows don't get installed in the switch.)
LbaaS is installing the OF Flows in opendaylight-inventory CONFIG data store with NodeId stored in OVSDB topology data tree, but openflow plugin doesn't have any node with this node Id as they are based on DPID.

Other NetVirt Provider Services map the OVSDB node Id to Southbound DPID, but LBaaS is not doing the mapping.

Fix is to get the corresponding DPID and create the Node Id that can be understood by the Openflow plugin and install the flow.

Change-Id: Ifd08df4d7729ebe6270db1afe239c3e8188dc229
Signed-off-by: Arun S. Yerra <arunsarat@yahoo.com>