Use MD-SAL events to react to OVSDB changes
Instead of just waiting after creating or deleting the OVSDB MD-SAL
node, wait for the corresponding event in the operational database.
This actually saves time in most cases (since the round-trip takes
less time than the static wait), and avoids failing tests
unnecessarily when the round-trip takes longer than expected. There's
an overall timeout to make sure we don't hang the test completely.

For now this only handles connectOvsdbNode() and
disconnectOvsdbNode(), but the stage is set for other MD-SAL
manipulations to be processed in the same way.

This isn't multi-threadable as-is.

Change-Id: If1279054f38192bed007d83f9fda981dc0c466f1
Signed-off-by: Stephen Kitt <skitt@redhat.com>