QoS Schema - fix type of QoS type
It looks like the 'type' field in the QoS table schema
was incorrectly coded in the Qos.java schema file
to be a Set<String>.
In the ovs schema, the type field for Qos is defined
as a simple string:

        "type": {
         "type": "string"},

Other string fields in the ovs schema that are
defined in the <table>.java files as Set<String>
appear in the schema as follows (for example,
ovs_version from Open_vSwitch table):

       "ovs_version": {
         "type": {"key": {"type": "string"},
                  "min": 0, "max": 1}},

With the QoS 'type' field defined as Set<String>,
the code gets an exception trying to cast
a String to a Set<String> when trying to get
the type out of a table update.  After this patch
is applied, the error is resolved.

Change-Id: Ibb42db03e71d8f1d96329b5942e7501069af20cb
Signed-off-by: Eric Multanen <eric.w.multanen@intel.com>