OVSDB library: Close conneciton when peer is down
In current implementation connection is not closed when peer is down
before it has chance to send TCP FIN packet. We have seen this issue
when peer is power off or crashed. With this fix, OVSDB server will
get IdleState.READER_IDLE event and send echo message to peer when
no channel read is performed for 60 seconds. OVSDB server will also
receive ReadTimeoutException and close the connection after 600 seconds.

Bug: 5506
Change-Id: I602b2ba3d34bc5e09546fe4e0584d8d0db42867e
Signed-off-by: Hsin-Yi Shen <syshen66@gmail.com>