it-utils should be only included for IT
The it-utils bundle was being included with the odl-ovsdb-openstack
feature. When the feature was installed karaf would error out
looking for org.junit - which is a dependency for it-utils since
that is a test bundle.

Fix: Create a IT feature and include the bundle there and use
that feature for the IT test.

Change-Id: I4fc0395ca177d98b8433857b384cddbc033057fb
Signed-off-by: Sam Hague <shague@redhat.com>