Update patch set 1
Patch Set 1:

> Curious why this is an issue in CSIT? Our IT works fine and the
 > build we made for OPNFV work fine.

CSIT requires /restconf/modules to be available in order to deem the
controller "UP".  this is somehow loaded when we load the -ui
feature set.

OPNFV is probably not hitting /restconf/modules and only looking in
topology for ovsdb:1 or netvirt:1 so they are ok.

this is a regression in the sense that it's different from what we 
had in our beryllium release.  you can see -ui is still there in the
now stale ovsdb repo.

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0