Bug 5781: Ping test is FAILED from DC-GW to Invisible IP Bug 5782: Subnet route is not getting programmed for a Network
For issue 5781:
The packets received by SubnetRouteHandler are all MPLS Label
wrapped if the packets are initiated by DC-GW.  As a result,
such packets are ignored by SubnetRoutePacketHandler.

This fix changes the SubnetRoute rule in Table 20 (LFIB Table)
to pop-off the mpls label before shipping packet to the
controller.

For issue 5782:
The race to fetch the RD by both VPNManager and SubnetRouteHandler
results in RD fetch failing in SubnetRouteHandler.  The fix is
for SubnetRouteHandler will now rely on VPNInstance config produced by
NeutronVpn rather than using vpn-instance-to-vpn-id config mapping
produced by VpnManager.

Change-Id: I73af156d0d55d7f79dca592c053f26398788b719
Signed-off-by: Suraj Ranjan <suraj.ranjan@ericsson.com>