Fix IPv6 Duplicate Address Detection Failure
In the current implementation, routermanager is responding with
a Neighbor Advt even for VM interfaces. Because of this, when a
VM is trying to assign an IPv6 address, the DAD phase fails and
VM is unable to configure the IPv6 address thinking that some
other VM is using the same IPv6 address. The following logs can
be seen on the VM syslog.

"IPv6 duplicate address 2001:1111::f816:3eff:fe7f:f643 detected!"

This patch addresses the issue by responding with an NA only to the
Router interfaces.

Change-Id: Ie743b5dd3fe80d4998484bc9faf20d81fd608a50
Signed-off-by: Sridhar Gaddam <sgaddam@redhat.com>