Bug 5860 Fix port event wrong ordering
If the southbound event arrives before the northbound event the
tenant network is not there yet and therefore the flows for
this port will not be created.

This patch add new logic to the PortHandler (northbound event)
to verify if the southbound event was already arrived and to
add the missing flows.

Change-Id: Id8ad03b6de8afff0b64ca3af3457281ddbaf0f74
Signed-off-by: oshvartz <oshvartz@redhat.com>