Update patch set 1
Patch Set 1: Code-Review-1

The neighbor solicitation resolution in the router manager plugin is for resolving ipv6 address from the external gateway router. For intra tenant NS resolution, the NS messages will be replied by the VMs themselves. So to resolve NS from external network, the router manager plugin must be aware of even the VM's interface ipv6 address. To avoid DAD failure when learning an ipv6 address for a VM interface, the NS messages from the intra tenant VMs has to be blocked in the Packet handler.

Patch-set: 1