Update patch set 1
Patch Set 1:

@Sridhar
Lets say we use static routing to manually program routes in the VMs and the external gateway router for the North-south traffic. In the ovs switch a default route via the gateway router and static mac programmed to reach the gateway router are added for sending the traffic out of the openstack env. For the return traffic from the ext gateway router, it needs to resolve the mac of the vm ip. This resolution will need to be done by the router manager plugin.

Patch-set: 1