Update patch set 1
Patch Set 1:

This optimization looks fine.

However, I have the following thoughts.

Currently we are programming a generic flow when a port is created on a node. This is causing all the NS/RS to be punted to the controller (irrespective of whether the network has IPv6 subnet or not). 

Shall we enhance the patch to check if IPv6 is enabled on the port, network is associated to the router, NS is for the router interface etc.

What are your thoughts on this?

Patch-set: 1