Update patch set 1
Patch Set 1:

yes, these 2 work items have to taken up certainly. But i don't want to combine them with this patch. 
How i perceive them is, it will be independent of these changes and this file in specific. So there is no reason to have them in a single patch.

Patch-set: 1