Update patch set 2
Patch Set 2:

Since the config yang doesn't provide a service and the createInstance is a no-op, we can remove the config yang, XML and Module class altogether. This is what I recently brought up on openflow-dev, although you were the only that responded :)  I also talked with Anil and Colin and they agreed we might as well clean up the config yang and bite the bullet now - we'll want to do it eventually anyway. I've pushed patches in openflowplugin to remove the config yang for ones that don't provide a service.

Patch-set: 2
Label: Verified=0