Update patch set 5
Patch Set 5:

> Thanks for the tip. Sometimes having the eyes to close to the
 > screen I can't see things anymore.
 > 
 > > You also have to create the nodeCacheListener bean.
 > 
 > Why would I need a been for this? Only PipelineOrchestratorImpl
 > implements this interface

I guess I wasn't clear on the purpose of that <service> element. Did you mean to advertise the PipelineOrchestratorImpl bean with the NodeCacheListener interface? If so then ref that bean or you can include the interface with other <service> element, i.e. you can advertise a bean with multiple interfaces in the same <service> tag. 

On a side, if that's the case, why does PipelineOrchestratorImpl need to advertise itself with the NodeCacheListener interface since it calls the NodeCacheManager directly (in that funky manner)?

Have I mentioned yet that the existing wiring code in this bundle is a bit of an ugly mess :)

Patch-set: 5