Bug 5939 - Communications through external gateway not working
The problem happens because the GatewayMacResolver is not resolving the
MAC address for the external gateway.  GatewayMacResolver is failing
because externalSubnet == null in triggerGatewayMacResolver.  This is
happening due to changes from [0] (a fix for BUG-5291).

In [0], a change was made to only set GatewayIP if port.getDeviceOwner()
== "network:router_interface"; however, in our case,
port.getDeviceOwner() == "network:router_gateway", so the Gateway IP
isn't set.

In this fix, I also set GatewayIP if port.getDeviceOwner() ==
"network:router_gateway".

[0] https://git.opendaylight.org/gerrit/#/c/37136/

Change-Id: I49b089900e4f85581beb8634580439dca2413ce5
Signed-off-by: Andre Fredette <afredette@redhat.com>