Update patch set 1
Patch Set 1:

@Sridhar
Should the router flag be set generically. Will there be a possibility of the ODL controller resolving the mac of one of the VMs   in that case, it is resolving the mac on behalf of the VM and source information in the NA is not of a router, so the router flag should not be set in this case right?

Patch-set: 1