Update patch set 7
Patch Set 7: Code-Review-1

Assume a situation where there are two network A and B. Once VM is connected to Network A and it was spawned on Compute 1. Both Network A and Network B are connected through router. Now you spawn another VM that is connected to network B and that VM gets spawned on compute -2. Now with the current logic, do you think this will work?

Patch-set: 7