Supporting RS packets when ipv6 subnets are configured with fixed ips.
When ipv6 Address mode and RA mode are not mentioned for an ipv6 subnet,
when responding to the RS messages, the packethandler is not able to access
these above mentioned modes of the router interface and crashes.
This is modified as when responding to the RS messages, with above mentioned
modes are not configured for a subnet, the RS message is still replied with
a RA message and the prefix associated with that subnet is not advertized
in the RA mesage.

Change-Id: Icc03a9385e13dfc5388a501e77e3d69d3ebb178e
Signed-off-by: kalaiselvik <Kalaiselvi_K@Dell.com>