Optimizing the icmpv6 flows programmed
The flow to punt the icmpv6 RS packets are programmed to a
node only when an ipv6 subnet is available on the node.
For responding to the Neighbor solicitation messages
from the external world, the packets are received only if
they are received on the patch-ext port.
For Neighbor solicitation within the VMs, the NS packets
are received on the associated tap interfaces, and these
packets are not punted to the controller.

Change-Id: Ic6082a666172138a765216c4fdcaa8d17dc9aa0e
Signed-off-by: kalaiselvik <Kalaiselvi_K@Dell.com>