Update patch set 3
Patch Set 3:

Alexis, can we take another pass at this cleanup or should we wait until blueprint is in?

Other question is we should break this patch into old netvirt and new netvirt. That way we can cherry-pick the old netvirt to beryllium. Or if it doesn't add value just keep it as one. Again, I think the blueprint stuff will have an impact and that is not on beryllium.

Patch-set: 3