Update patch set 3
Patch Set 3: Code-Review-1

added Aswin and Isaku to the review to help with Prashanth's question.

But, the reason the originalData is null is because of the path this method is called. The method with the exception is expecting to be called from an mdsal update and called in the NeutronPortchangeListener#updatePort() path. In that case the originalPort is set and the method works fine. But the failure here, the handleIntrrfaceEvent() is being called from the SouthBoundHandler. In that path the originalPort is not set.

So this patch here should keep the checks as is, but we should also ensure that the originalport is set before this method is called. But we normally get the value from the mdsal dataChange event coming in. We woudl need to see if just using the same port as the original port would be fine. it may also be that is doesn't matter in this code path what value is used since the updatePort(0 path should cath what we want.

Aswin, can you weigh in on this?

Patch-set: 3
Label: Code-Review=-1