Update patch set 3
Patch Set 3:

The update port is set only when the port update event is received from neutron northbound . In case of other events like northbound create or southbound create/update it will be null. So it is safe to return false in this.processPortSecurityEnableUpdated(neutronPort) when it the original port is null, since we are sure it not an update from northbound(the only way we can toggle portSecurityEnabled field).

Setting the current port as the original port is safe as well as there is further check for the delta of the particular attribute the function is interested in. It may help in future when this original port is consumed by any other service for the checking the delta from not being hit with a NOP, may be the better option.

Patch-set: 3
Reviewer: Gerrit User 4269 <4269@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0