Update patch set 1
Patch Set 1:

Changes look fine but I had a generic question about our approach. Looks like we're just going to augment acl model with security groups and dump them as is and not using the existing fields provided by acl model. Correct? If yes, why not reuse existing fields and translate security groups and rules to ACL model? I realize that would take time and might not be feasible within Boron timeframe and we could still do it later as it is an implementation detail, but was wondering if any specific reason/limitation that made us not consider it.

Patch-set: 1