Update patch set 1
Patch Set 1:

We're not really adding many extra fields, mainly Neturon SG specific uuids and direction. The uuids confused me a bit for how we're planning to use it. So my question was if we're still going to translate SG matches/rules to ACL matches and fill in this additional info for reference, or we'll pull security group information from NeutronSGs even in ACL service impl code. I thought it was the former but wanted to be sure.

Patch-set: 1