Update patch set 2
Patch Set 2:

(5 comments)

i think that both the fix and the unit-tests should be on the same merge otherwise in case of a revert we will expect the one that do the revert to be aware of the additional merge.
in my opinion it is much easier to understand the fix by going over the unit-tests. so it will be easier doing that if both changes are in one patch

Patch-set: 2
Label: Verified=0