Optimizing the icmpv6 flows programmed
The flow to punt the icmpv6 RS packets are programmed to a
node only when an ipv6 subnet is available on the node.
The flow to punt icmpv6 NS packets are segregated as to punt
only the NS packets received for the ipv6 router interfaces
and for the NS packets received from the external network.

Other minor changes to accomodate the changes in the neutron
and the openflowplugin.

Change-Id: Ic6082a666172138a765216c4fdcaa8d17dc9aa0e
Signed-off-by: kalaiselvik <Kalaiselvi_K@Dell.com>