Auto-creation of bridges in vpnservice
When a new OVSDB node is detected, the
event is caught and we make sure that
br-int and br-ex (if l3 is configured)
exist. Note that br-ex is not technically
required and once devstack is modified
to not check for it, we should go back
in to this code and no longer create it.

This code add dependencies to the
following ovsdb project utility bundles
and jars:
- southbound-api
- utils.mdsal-utils
- utils.southbound-utils
- utils.config

Change-Id: I852c2b619c99dc2748808efe27d8d24487e0cee7
Signed-off-by: Josh <jhershbe@redhat.com>