Update patch set 1
Patch Set 1: Code-Review-1

This is expected that the NodeCacheManager could be instantiated after the PipeLineOrchestrator. That is why setDependencies is two different methods. The first sets dependencies that are know to be allocated since they were concrete in the activator. The other is for listening for when the dependency is allocated. This patch would mix that a bit and I don't think really solves the problem.

What is the scenario where the NPE hits? I could see Of13Provider coming up and triggering the netvirt to start the provider and dataChangeListener. And the nodeCacheManager might not have been set. We haven't seen that. But if it does happen then I think we need to hold off on the start event until all the dependencies are set.

Patch-set: 1
Reviewer: Gerrit User 773 <773@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1