Update patch set 2
Patch Set 2:

Hello Daya/Srikanth, in one of the previous meetings, we discussed about the approach to be taken for Neutron Router/Subnet/Port listeners. 
I'm not sure if we have come to some conclusion. 

Basically, I see that we have two options.

1. To listen to the events directly from genius mdsal utils (like NeutronRouterAdded, NeutronSubnetAdded etc). This patch has these listeners (f.e, NeutronPortChangeListener.java)
2. To register for the notifications from NeutronVPN (i.e., L3 service in NetVirt) as per this yang [*]. Where we can get notified for events like onPortAddedToSubnet, onRouterAddedToVPN etc) and use rpc methods to get the associated info.

Can you please share your thoughts on the approach to be taken. If its 2 above, my followup question is "Can we extend the YANG [*] with some new notifications/rpc methods (if required) for IPv6 service"?

Awaiting your response. Thanks.   

[*] https://github.com/opendaylight/netvirt/blob/master/vpnservice/neutronvpn/neutronvpn-api/src/main/yang/neutronvpn.yang#L372

Patch-set: 2