Update patch set 2
Patch Set 2: Code-Review+2

Michael, this looks good. Would it be cleaner in another patch to make a Table class holding all these constants (or enum)? These table id's are consistent across the whole of netvirt-vpnservice. It might even make sense to be in genius since that is the table dispatcher. But, it seems the constants (or enums) should be centralized.

Patch-set: 2
Reviewer: Gerrit User 773 <773@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+2