Uncomment install/removeFlow
I cannot write a test for aclservice if actually it does NOTHING. The
current implementation just pushes some bits from here to there
internally, but ends up doing nothing.  I'm working on a showcase test
that will demonstrate how I would test a module such as aclservice. It
will show how to test the EFFECT of an API, and for that, it must DO
something real... ;-)

Change-Id: I9b240076820bc6b49b4891caf669a43f24335d3b
Signed-off-by: Michael Vorburger <vorburger@redhat.com>