Create physnet bridges+ports and patch to br-int
NB: This commit depends on changes in the ovsdb
project: https://git.opendaylight.org/gerrit/#/c/41946/

For each physical net in Open_vSwitch other_config provider_networks
we create a bridge called br-ex-<interface_name>
and the interface is placed in that bridge. The bridge
is then patched to br-int (old br-ex is not used).

This is all triggered from ElanOvsdbNodeListever.
The creation of the various bridges and ports is tied
together asynchronously and we wait for notification
from OPERATIONAL that the correct bridge exists before
moving to the next step.

Step 1: OVSDB node notification
 -> create br-int and br-ex
Step 2: br-int created notification
 -> create the bridges for each external net interface
Step 3: br-ex-<iface> created notification
 -> add the ports and patch to br-int

Change-Id: Ibf6851d420c8197e11a6503b950fa381b829499c
Signed-off-by: Josh <jhershbe@redhat.com>