Update patch set 3
Patch Set 3:

> I can see new trailing spaces..
 > Not sure why the checkstyle plugin didn't catch it.

because they are in src/test and not src/main.  I'll be adding that under https://git.opendaylight.org/gerrit/#/c/40957/.  (The problem is that existing projects with enforcement need to be cleaned up first; that takes some time. Help welcome; email me if you would have time to help me with moving forward with that c/40957).

Patch-set: 3
Label: Verified=0