Update patch set 1
Patch Set 1:

Configuration of Transport Zones, one per network.
The tunnel is configured on VM up (Neutron port full creation), as it was done in the old netvirt.
The trigger to add VTEP (or create new Transport Zone) is addition of Interface state in OPERATIONAL data stores.

Patch-set: 1
Label: Verified=0