Update patch set 20
Patch Set 20:

Much of this code was refactored in Josh's review which was built on top of this one. The only thing really needed by the ELAN service is the external port name and this is still WIP on review #41925.
I can create a new bundle in this review and add a placeholder method to get the external port name.

Patch-set: 20