Write netvirt:1 to mdsal when netvirt is ready
This new module will write netvirt:1 to mdsal when all the
netvirt services have started. There are 10 services that
are required. Some services are in the Service Registry and
can be referenced in the blueprint.xml. Other services can
only be found via the config subssystem beans. Once we move
the rest of netvirt to blueprint we can deprecate the config
related code.

Change-Id: I3841609c18ba65463d8e7ee07304fa0210b193f7
Signed-off-by: Sam Hague <shague@redhat.com>