TR : HU97823 Sometimes Table 0 not populated
Issue : Sometimes Table 0 not properly populated for Trunk Port

Analysis :

Parent id is coming wrong in the InterfaceConfigDS. But its correct in the
Neutronport config DS. So Interface manager is not adding the states
correctly in the OVS
resulting in table 0 miss population.
On analyzing neutronvpnmanager, Parent-interface is not correctly
populated for Subports in method updateInterface() of
NeutronPortChangeListener for port VIF type updated scenarios.

Fix :

updateInterface() method of NeutronPortChangeListener is modified to
handle trunk
interface case also

Change-Id: If4bfae25d9fbc7ab88546ff034582aece22b1b1c
Signed-off-by: Vyshakh Krishnan CH <vyshakh.krishnan.c.h@ericsson.com>