Update patch set 5
Patch Set 5:

(3 comments)

@Srikanth,

After an unsolicited RA msg is transmitted in the transmitUnsolicitedRA(), the setPeriodicTransmissionTimeout() is invoked to add the periodicTransmission timeout object to the timer wheel. So on expiry of the timeout object, again the cycle of sending RA and adding the timeout object continues. Hope this clarifies.

Patch-set: 5
Reviewer: Gerrit User 3374 <3374@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0