Update patch set 1
Patch Set 1:

My point is interfaces does not get added/removed or vm gets migrated very frequently. But still we are doing the read operation for every timer expiry seems to be a concern.
Now when we have made the periodic interval configurable, and the user configures very small interval, then the no. of times the read is done will increase.
If might be better if the local cache is maintained in a central location so it can be accessible to all.
Or we can restrict the no. of reads done for every fixed time period as data will not be modified frequently with respect to the timer interval duration.

Patch-set: 1