Update patch set 1
Patch Set 1:

This patch is not a critical patch as its only an optimization of the existing implementation.
So, lets continue to have an healthy discussion on this, as it helps us (/others) in future designs.
BTW, I'm happy to amend the patch (with some local cache) if its the general recommendation.

In genius-dev mailing list for some similar requirement, the suggestion was to use datastore [1]. 
[1] https://lists.opendaylight.org/pipermail/genius-dev/2016-July/000175.html

It would be good to hear the views of experts in this area for their suggestion.

PSB my comments to your query.

Kalai>> Now when we have made the periodic interval configurable, and the user configures very small interval, then the no. of times the read is done will increase.
[Sridhar] Ideally i would not expect the user to choose a smaller interval, unless they have a valid reason. Anyways, I think this is not the crux of the discussion, so we can discuss this during our weekly call.

Kalai>> If might be better if the local cache is maintained in a central location so it can be accessible to all.
[Sridhar] Agreed. Is this maintained somewhere?

Kalai>> Or we can restrict the no. of reads done for every fixed time period as data will not be modified frequently with respect to the timer interval duration.
[Sridhar] hmmm...personally I would like to avoid this as it leads to race conditions.

Patch-set: 1