Fixed the NAT issue SNAT flows not getting configured
The model NeutronRouterDpns, which is one of the main models used
by SNAT, is not getting populated because of the dependent model
RouterInterfaces being empty. RouterInterfaces model is empty because
of the race condition b/n the RouterInterfacesListner thread
which is trying to read interface state before the thread which
updates the IetfInterfaces model. Hence in RouterInterfacesListener, we
have removed the check of operational state of Interfaces before populating the
RouterInterfaces model. Eventually InterfacesStateChangeListener will
read RouterInterfaces only when ietf-interface is updated and populates
the NeutronRouterDpns model

Change-Id: I2ffaa9d686d213bcd8187ac9e41c5ad2ae5945d8
Signed-off-by: Satish Dutt <satish.dutt@ericsson.com>