Update patch set 1
Patch Set 1:

While working on Optimal flow programming for IPv6 service binding using elanTag [1], there was a need to maintain a cache of dpnIds <--> List[ports] per network. We can use the same cache for this patch and avoid calling dataStore. Thanks for all your comments. Currently the patch [1] is failing in Jenkins for a different issue. I'll update this patch soon by re-basing it on top of patch[1].

[1] https://git.opendaylight.org/gerrit/#/c/43477/

Patch-set: 1