Update patch set 19
Patch Set 19: Code-Review-1

Remove elanmservice-impl from the readyMap in ConfigStateManager. Also remove the call to start ConfigStateManager and call setReady(true) in the StateManager.start() since there are no config style beans to look for anymore. The cloudchain bundle has not been converted to blueprint yet so we should keep the ConfigStateManager around in case we need it for that until it is converted.

Patch-set: 19
Label: Code-Review=-1