Update patch set 5
Patch Set 5:

There's a known issue that the removal of the group is failing because of its type.
This issue is in openflowplugin-impl - no matter what type is set it will use the type the group had when it was created and in our case indirection. The OVS will return an invalid group error and will not remove the group when indirection type is used.
I've sent to the openflowplugin mailing list a message about it.

Patch-set: 5
Reviewer: Gerrit User 6248 <6248@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0