Update patch set 1
Patch Set 1:

The following lead me to believe the intention is otherwise:
1) Why check the operational then?
2) The other way around, if it is not visible in the operational for whatever reason, why re-add it to the configuration if it is already there?
3) Then there is the fact that this could be overriding the configuration already done by an external entity. Elan manager previously did not add any bridges and it is easier to understand if this change was introduced in a less intrusive way. 
4) Although it is not fortright clear, I still take from the comment of the method that the intention is to add the bridge if it is neither in configuration or operational. 

Anyway, it would be nice to be sure about this because it's going to impact people that were previously working with elan manager but adding the bridges themselves. On a real neutron deployment, isn't this done by the neutron agent?

Patch-set: 1
Reviewer: Gerrit User 6230 <6230@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0