Update patch set 1
Patch Set 1:

One reason to check the operational is that if for whatever reason the connection was not made and therefor the node didn't wind up in OP rewriting to config could trigger a retry. This is actually quite old code that we dragged over from the old netvirt and I would be reluctant to touch it at all especially the change proposed here which *narrows* the condition under which the bridge is created.

Patch-set: 1