Update patch set 1
Patch Set 1: Code-Review-1

It may not be that the check is wrong, but the southboundUtils.addBridge is wrong. For the scenario where the bridge is already created externally, then the ovsdbNode update comes in with the bridge so the first part of the check !isBridgeOnNode() returns false (!true) but then the getBridgeFromConfig()==null returns true since it won't be in config. So we do the addBridge() - and that does a put() with totally new bridgeAugmentation. What should happen is the method should take the existing bridgeAugmentation and just update the controller and merge that back to config so that it doesn't change the dpid.

Patch-set: 1