Update patch set 1
Patch Set 1:

I would still need to make a decision about the dptype and mac address that are being passed to it. 
Would it be reasonable to use the other version of SouthBoundUtils::addBridge #526 and handle that from the ElanBridgeManager: if there is already a bridge augmentation  for that bridge, just extract the info from there and call the other addBridge method?

Patch-set: 1