Update patch set 1
Patch Set 1:

Michael, this looks pretty cool. Much less error-prone.

Are there annotations for finding different versions of services? For example, the DataBroker has a few different impls, of which we use the odl-default, but there are other's like ping-pong.

There are also notification providers for things like the packetHandlers where the service implements an interface.

Patch-set: 1
Label: Verified=0