BUG-6495 BC Grp wrong for E/W VLAN provider net
As per the explanation in the bug, this group
is configured twice and one of these times is
broken for VLAN and only works for VxLan.
Easy way to solve this it to only call the second
method when the network type is VxLan. That's
what this patch does. Would be better if this
group was configured in one place, coherently.

Change-Id: Ia5d8f003d2f591b6a3d366928683c9e1a3515eb6
Signed-off-by: Josh <jhershbe@redhat.com>