Update patch set 1
Patch Set 1: Code-Review-1

> That is great!
 > Why haven't you changed all the classes to use this plugin?

This is a proof-of-concept for the time being. See the discussion on https://git.opendaylight.org/gerrit/44502 — there are some issues we need to fix before we can actually use something like this unfortunately!

In particular we need to be absolutely sure that all the contexts are loaded, from both org/opendaylight/blueprint and OSGI-INF/blueprint (hence my -1 for now).

 > Is it supported in maven 3.3.9?

It's supported by an additional Maven plugin, as long as Maven is new enough to run that plugin (3.0.5 apparently) it will work.

Patch-set: 1
Label: Code-Review=-1